move composition playback into custom hook
